rename chapter7 to chapter6
