+++ src/main/java/org/zkoss/essentials/chapter5/MenuNode.java
