+++ src/main/java/org/zkoss/essentials/chapter5/MenuViewModel.java
