+++ src/main/java/org/zkoss/essentials/chapter6/ajaxbased/SidebarAjaxbasedController.java
@@ -6,7 +6,7 @@
 
 Copyright (C) 2012 Potix Corporation. All Rights Reserved.
 */
-package org.zkoss.essentials.chapter7.pagebased;
+package org.zkoss.essentials.chapter6.ajaxbased;
 
 import org.zkoss.essentials.services.SidebarPage;
 import org.zkoss.essentials.services.SidebarPageConfig;
@@ -17,21 +17,23 @@ import org.zkoss.zk.ui.event.EventListener;
 import org.zkoss.zk.ui.event.Events;
 import org.zkoss.zk.ui.event.SerializableEventListener;
 import org.zkoss.zk.ui.select.SelectorComposer;
+import org.zkoss.zk.ui.select.Selectors;
 import org.zkoss.zk.ui.select.annotation.Wire;
 import org.zkoss.zul.Grid;
 import org.zkoss.zul.Image;
+import org.zkoss.zul.Include;
 import org.zkoss.zul.Label;
 import org.zkoss.zul.Row;
 import org.zkoss.zul.Rows;
 
-public class SidebarPagebasedController extends SelectorComposer<Component>{
+public class SidebarAjaxbasedController extends SelectorComposer<Component>{
 
 	private static final long serialVersionUID = 1L;
 	@Wire
 	Grid fnList;
 	
 	//wire service
-	SidebarPageConfig pageConfig = new SidebarPageConfigPagebasedImpl();
+	SidebarPageConfig pageConfig = new SidebarPageConfigAjaxBasedImpl();
 	
 	@Override
 	public void doAfterCompose(Component comp) throws Exception{
@@ -43,10 +45,10 @@ public class SidebarPagebasedController extends SelectorComposer<Component>{
 		for(SidebarPage page:pageConfig.getPages()){
 			Row row = constructSidebarRow(page.getName(),page.getLabel(),page.getIconUri(),page.getUri());
 			rows.appendChild(row);
-		}
+		}		
 	}
 
-	private Row constructSidebarRow(String name,String label, String imageSrc, final String locationUri) {
+	private Row constructSidebarRow(final String name,String label, String imageSrc, final String locationUri) {
 		
 		//construct component and hierarchy
 		Row row = new Row();
@@ -58,18 +60,33 @@ public class SidebarPagebasedController extends SelectorComposer<Component>{
 		
 		//set style attribute
 		row.setSclass("sidebar-fn");
-			
-		EventListener<Event> actionListener = new SerializableEventListener<Event>() {
+		
+		//new and register listener for events
+		EventListener<Event> onActionListener = new SerializableEventListener<Event>(){
 			private static final long serialVersionUID = 1L;
 
 			public void onEvent(Event event) throws Exception {
 				//redirect current url to new location
-				Executions.getCurrent().sendRedirect(locationUri);
+				if(locationUri.startsWith("http")){
+					//open a new browser tab
+					Executions.getCurrent().sendRedirect(locationUri);
+				}else{
+					//use iterable to find the first include only
+					Include include = (Include)Selectors.iterable(fnList.getPage(), "#mainInclude")
+							.iterator().next();
+					include.setSrc(locationUri);
+					
+					//advance bookmark control, 
+					//bookmark with a prefix
+					if(name!=null){
+						getPage().getDesktop().setBookmark("p_"+name);
+					}
+				}
 			}
-		};
-		
-		row.addEventListener(Events.ON_CLICK, actionListener);
+		};		
+		row.addEventListener(Events.ON_CLICK, onActionListener);
 
 		return row;
 	}
+	
 }
