+++ src/main/java/org/zkoss/essentials/chapter6/ajaxbased/mvvm/SidebarAjaxbasedViewModel.java
@@ -6,10 +6,11 @@
 
 Copyright (C) 2012 Potix Corporation. All Rights Reserved.
 */
-package org.zkoss.essentials.chapter7.ajaxbased.mvvm;
+package org.zkoss.essentials.chapter6.ajaxbased.mvvm;
 
 import java.util.List;
 
+import org.zkoss.essentials.chapter6.ajaxbased.SidebarPageConfigAjaxBasedImpl;
 import org.zkoss.essentials.services.SidebarPage;
 import org.zkoss.essentials.services.SidebarPageConfig;
