+++ src/main/webapp/chapter6/ajaxbased/index.zul
@@ -1,14 +1,14 @@
 <?link rel="stylesheet" type="text/css" href="/style.css"?>
 <zk>
-	<borderlayout hflex="1" vflex="1" apply="org.zkoss.essentials.chapter7.ajaxbased.BookmarkChangeController">
+	<borderlayout hflex="1" vflex="1" apply="org.zkoss.essentials.chapter6.ajaxbased.BookmarkChangeController">
 		<north height="100px" border="none" >
 			<apply templateURI="/chapter1/banner.zul"/>
 		</north>
 		<west width="260px" border="none" collapsible="true" splittable="true" minsize="300">
-			<apply templateURI="/chapter7/ajaxbased/sidebar.zul"/>
+			<apply templateURI="/chapter6/ajaxbased/sidebar.zul"/>
 		</west>
 		<center id="mainContent" autoscroll="true" border="none" self="@insert(content)">
-			<apply templateURI="/chapter7/ajaxbased/mainContent.zul"/>
+			<apply templateURI="/chapter6/ajaxbased/mainContent.zul"/>
 		</center>
 		<south height="50px" border="none">
 			<apply templateURI="/chapter1/footer.zul"/>
