+++ src/main/webapp/chapter6/ajaxbased/mainContent.zul
@@ -1,0 +1,3 @@
+<zk>
+	<include id="mainInclude"  src="/chapter6/ajaxbased/home.zul"/>
+</zk>
