+++ src/main/webapp/chapter6/ajaxbased/sidebar.zul
@@ -1,6 +1,6 @@
 <grid hflex="1" vflex="1" sclass="sidebar" 
 	id="fnList" 
-	apply="org.zkoss.essentials.chapter7.ajaxbased.SidebarAjaxbasedController">
+	apply="org.zkoss.essentials.chapter6.ajaxbased.SidebarAjaxbasedController">
 	<columns>
 		<column width="36px"/>
 		<column/>
