+++ src/main/webapp/chapter6/ajaxbased_mvvm/home.zul
