+++ src/main/webapp/chapter6/ajaxbased_mvvm/index.zul
@@ -1,15 +1,15 @@
 <?link rel="stylesheet" type="text/css" href="/style.css"?>
 <zk>
-	<borderlayout hflex="1" vflex="1" viewModel="@id('vm') @init('org.zkoss.essentials.chapter7.ajaxbased.mvvm.BookmarkChangeViewModel')" 
+	<borderlayout hflex="1" vflex="1" viewModel="@id('vm') @init('org.zkoss.essentials.chapter6.ajaxbased.mvvm.BookmarkChangeViewModel')" 
  			onBookmarkChange="@command('onBookmarkNavigate', bookmark=event.bookmark)"> 
 		<north height="100px" border="none" >
 			<apply templateURI="/chapter1/banner.zul"/>
 		</north>
 		<west width="260px" border="none" collapsible="true" splittable="true" minsize="300">
-			<apply templateURI="/chapter7/ajaxbased_mvvm/sidebar.zul"/>
+			<apply templateURI="/chapter6/ajaxbased_mvvm/sidebar.zul"/>
 		</west>
 		<center id="mainContent" autoscroll="true" border="none" >
-			<apply templateURI="/chapter7/ajaxbased_mvvm/mainContent.zul"/>
+			<apply templateURI="/chapter6/ajaxbased_mvvm/mainContent.zul"/>
 		</center>
 		<south height="50px" border="none">
 			<apply templateURI="/chapter1/footer.zul"/>
