+++ src/main/webapp/chapter6/ajaxbased_mvvm/mainContent.zul
@@ -1,0 +1,5 @@
+<zk>
+	<include id="mainInclude"
+		viewModel="@id('vm') @init('org.zkoss.essentials.chapter6.ajaxbased.mvvm.NavigationViewModel')" 
+		src="@load(vm.includeSrc)" />
+</zk>
