+++ src/main/webapp/chapter6/pagebased/index-todolist-mvvm.zul
@@ -1,7 +1,7 @@
 <?link rel="stylesheet" type="text/css" href="/style.css"?>
 <zk>
 <div height="100%">
-	<template name="layout" src="/chapter7/pagebased/layout/template.zul"/>
+	<template name="layout" src="/chapter6/pagebased/layout/template.zul"/>
 	<apply template="layout"/>
 	<template name="center" src="/chapter4/todolist-mvvm.zul"/>
 </div>
