+++ src/main/webapp/chapter6/pagebased/layout/sidebar.zul
@@ -1,6 +1,6 @@
 <grid hflex="1" vflex="1" sclass="sidebar" 
 	id="fnList" 
-	apply="org.zkoss.essentials.chapter7.pagebased.SidebarPagebasedController">
+	apply="org.zkoss.essentials.chapter6.pagebased.SidebarPagebasedController">
 	<columns>
 		<column width="36px"/>
 		<column/>
