--- src/main/webapp/chapter7/ajaxbased/mainContent.zul
@@ -1,3 +1,0 @@
-<zk>
-	<include id="mainInclude"  src="/chapter7/ajaxbased/home.zul"/>
-</zk>
