--- src/main/webapp/chapter7/ajaxbased_mvvm/mainContent.zul
@@ -1,5 +1,0 @@
-<zk>
-	<include id="mainInclude"
-		viewModel="@id('vm') @init('org.zkoss.essentials.chapter7.ajaxbased.mvvm.NavigationViewModel')" 
-		src="@load(vm.includeSrc)" />
-</zk>
