--- src/main/webapp/chapter7/pagebased/index.zul
@@ -1,8 +1,0 @@
-<?link rel="stylesheet" type="text/css" href="/style.css"?>
-<zk>
-<div height="100%">
-	<template name="layout" src="/chapter7/pagebased/layout/template.zul"/>
-	<apply template="layout"/>
-	<template name="center" src="/chapter7/pagebased/home.zul"/>
-</div>
-</zk>
