+++ src/main/webapp/chapter8/layout/template.zul
@@ -5,7 +5,7 @@
 			<include src="/chapter8/layout/banner.zul"/>
 		</north>
 		<west width="260px" border="none" collapsible="true" splittable="true" minsize="300">
-			<include src="/chapter7/ajaxbased/layout/sidebar.zul"/>
+			<include src="/chapter6/ajaxbased/layout/sidebar.zul"/>
 		</west>
 		<center id="mainContent" autoscroll="true" border="none" self="@insert(content)">
 			<!-- the main content will be insert to here -->
